requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
fastapi==0.118.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.11.0
loguru==0.7.3
motor==3.7.1
multidict==6.6.4
openai==2.1.0
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.15.2
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.20.1
pydantic_settings
APScheduler